Reorganize apiserver routes into submodules
